@@ -1,26 +1,27 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Microsoft.AspNetCore.Builder;
+using QA.DotNetCore.Engine.CacheTags;
+using QA.DotNetCore.Engine.CacheTags.Configuration;
+using QA.WidgetPlatform.Api.Application.Middleware;
+using QA.WidgetPlatform.Api.Infrastructure;
 
-namespace QA.WidgetPlatform.Api
+var builder = WebApplication.CreateBuilder(args);
+builder.Services.ConfigureBaseServices(builder.Configuration);
+    
+var app = builder.Build();
+
+app.UseMiddleware<StatusCodeExceptionHandlerMiddleware>();
+//мидлвара для инвалидации кештегов
+//необходимо, чтобы было подключено services.AddCacheTagServices
+app.UseCacheTagsInvalidation(trackers =>
 {
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+    //регистрация одного или нескольких ICacheTagTracker
+    //QpContentCacheTracker - уже реализованный ICacheTagTracker, который работает на базе механизма CONTENT_MODIFICATION из QP
+    trackers.Register<QpContentCacheTracker>();
+});
+app.UseRouting();
+app.UseAuthorization();
+app.UseSwaggerUI();
+app.MapControllers();
+app.MapSwagger();
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
-    }
-}
+app.Run();
