@@ -1,9 +1,11 @@
 using System;
 using System.IO;
+using System.Text.Json.Serialization;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.OpenApi.Models;
+using QA.DotNetCore.Engine.CacheTags.Configuration;
 using QA.DotNetCore.Engine.Persistent.Interfaces.Settings;
 using QA.DotNetCore.Engine.QpData.Configuration;
 using QA.WidgetPlatform.Api.Services;
@@ -17,7 +19,7 @@ namespace QA.WidgetPlatform.Api.Infrastructure
         {
             services.AddControllers().AddJsonOptions(options =>
             {
-                options.JsonSerializerOptions.IgnoreNullValues = true;
+                options.JsonSerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull;
             });
 
             services.AddSwaggerGen(options =>
@@ -31,9 +33,28 @@ namespace QA.WidgetPlatform.Api.Infrastructure
             });
 
             services.AddMemoryCache();
+            var qpSettings = configuration.GetSection("QpSettings").Get<QpSettings>();
             services.AddSiteStructure(options =>
             {
-                options.UseQpSettings(configuration.GetSection("QpSettings").Get<QpSettings>());
+                options.UseQpSettings(qpSettings);
+            });
+
+            //подключение сервисов для работы кештегов
+            services.AddCacheTagServices(options =>
+            {
+                //настройка стратегии инвалидации по кештегам
+                if (qpSettings.IsStage)
+                {
+                    //при каждом запросе запускать все зарегистрированные ICacheTagTracker,
+                    //чтобы получить все теги по которым нужно сбросить кеш, и сбросить его
+                    options.InvalidateByMiddleware(@"^.*\/(__webpack.*|.+\.[a-zA-Z0-9]+)$");//отсекаем левые запросы для статики (для каждого сайта может настраиваться индивидуально)
+                }
+                else
+                {
+                    //по таймеру запускать все зарегистрированные ICacheTagTracker,
+                    //чтобы получить все теги по которым нужно сбросить кеш, и сбросить его
+                    options.InvalidateByTimer( TimeSpan.FromSeconds(30));
+                }
             });
             services.AddScoped<ISiteStructureService, SiteStructureService>();
             services.TryAddSingleton<ITargetingFiltersFactory, EmptyTargetingFiltersFactory>();
