@@ -1,8 +1,17 @@
-﻿namespace QA.WidgetPlatform.Api.Models
+﻿using System.Text.Json.Serialization;
+
+namespace QA.WidgetPlatform.Api.Models
 {
     public class FieldInfo
     {
-        public string Type { get; set; }
-        public object Value { get; set; }
+        public FieldInfo(string type, object value)
+        {
+            Type = type;
+            Value = value;
+        }
+
+        [JsonIgnore]
+        public string Type { get; }
+        public object Value { get; }
     }
 }
